chore: fmt

+++ contracts/src/test/7683/SolverRepaymentBatchingTest.t.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.25;
 
-import {T1MurkyMerkle} from "../utils/T1MurkyMerkle.sol";
+import { T1MurkyMerkle } from "../utils/T1MurkyMerkle.sol";
 
 import { IT1ERC7683 } from "../../../src/interfaces/IT1ERC7683.sol";
 
@@ -17,7 +17,6 @@ struct MerkleLeafHelper {
 }
 
 contract SolverRepaymentBatchingTest is T1XChainReaderTest {
-
     T1MurkyMerkle internal tree = new T1MurkyMerkle();
 
     uint256 internal intentCount = 20;
@@ -28,7 +27,7 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
 
         bytes32[] memory treeLeaves = new bytes32[](intentCount);
         MerkleLeafHelper[] memory helpers = new MerkleLeafHelper[](intentCount);
-        for (uint i = 0; i < intentCount; i++) {
+        for (uint256 i = 0; i < intentCount; i++) {
             helpers[i] = _openFillOrder_and_generateMerkleLeaf(kakaroto, vegeta, amounts[i]);
             treeLeaves[i] = helpers[i].treeLeaf;
             summedAmounts += amounts[i];
@@ -38,7 +37,7 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
         originReader.commitProofOfReadRoot(batchIndex, root);
 
         bytes[] memory encodedProofs = new bytes[](intentCount);
-        for (uint i = 0; i < intentCount; i++) {
+        for (uint256 i = 0; i < intentCount; i++) {
             bytes memory flattenedProof = _flattenProof(tree.getProof(treeLeaves, i));
             encodedProofs[i] = abi.encode(batchIndex, helpers[i].requestId, i, helpers[i].result, flattenedProof);
         }
@@ -46,11 +45,13 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
         uint256 balanceSolverBeforeSettle = inputToken.balanceOf(address(vegeta));
 
         uint256 gasBefore = gasleft();
-        for (uint i = 0; i < intentCount; i++) {
+        for (uint256 i = 0; i < intentCount; i++) {
             l1T1ERC7683.handleReadResultWithProof(encodedProofs[i]);
         }
         uint256 gasAfter = gasleft();
-        console.log("Gas used for handling proofs in test_ERC7683BatchSolverRepayment_noBatching:", gasBefore - gasAfter);
+        console.log(
+            "Gas used for handling proofs in test_ERC7683BatchSolverRepayment_noBatching:", gasBefore - gasAfter
+        );
 
         uint256 balanceSolverAfterSettle = inputToken.balanceOf(address(vegeta));
 
@@ -60,8 +61,12 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
             "vegeta balance increased by batch inputs amount"
         );
 
-        for (uint i = 0; i < intentCount; i++) {
-            assertEq(uint8(l1T1ERC7683.orderStatus(helpers[i].orderId)), uint8(IT1ERC7683.Status.SETTLED), "Order should be settled");
+        for (uint256 i = 0; i < intentCount; i++) {
+            assertEq(
+                uint8(l1T1ERC7683.orderStatus(helpers[i].orderId)),
+                uint8(IT1ERC7683.Status.SETTLED),
+                "Order should be settled"
+            );
         }
     }
 
@@ -71,7 +76,7 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
 
         bytes32[] memory treeLeaves = new bytes32[](intentCount);
         MerkleLeafHelper[] memory helpers = new MerkleLeafHelper[](intentCount);
-        for (uint i = 0; i < intentCount; i++) {
+        for (uint256 i = 0; i < intentCount; i++) {
             helpers[i] = _openFillOrder_and_generateMerkleLeaf(kakaroto, vegeta, amounts[i]);
             treeLeaves[i] = helpers[i].treeLeaf;
             summedAmounts += amounts[i];
@@ -81,7 +86,7 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
         originReader.commitProofOfReadRoot(batchIndex, root);
 
         bytes[] memory encodedProofs = new bytes[](intentCount);
-        for (uint i = 0; i < intentCount; i++) {
+        for (uint256 i = 0; i < intentCount; i++) {
             bytes memory flattenedProof = _flattenProof(tree.getProof(treeLeaves, i));
             encodedProofs[i] = abi.encode(batchIndex, helpers[i].requestId, i, helpers[i].result, flattenedProof);
         }
@@ -91,7 +96,10 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
         uint256 gasBefore = gasleft();
         l1T1ERC7683.handleBatchOfReadResultsWithProofs(encodedProofs);
         uint256 gasAfter = gasleft();
-        console.log("Gas used for handling proofs in test_ERC7683BatchSolverRepayment_sameSolver_withBatching:", gasBefore - gasAfter);
+        console.log(
+            "Gas used for handling proofs in test_ERC7683BatchSolverRepayment_sameSolver_withBatching:",
+            gasBefore - gasAfter
+        );
 
         uint256 balanceSolverAfterSettle = inputToken.balanceOf(address(vegeta));
 
@@ -101,8 +109,12 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
             "vegeta balance increased by batch inputs amount"
         );
 
-        for (uint i = 0; i < intentCount; i++) {
-            assertEq(uint8(l1T1ERC7683.orderStatus(helpers[i].orderId)), uint8(IT1ERC7683.Status.SETTLED), "Order should be settled");
+        for (uint256 i = 0; i < intentCount; i++) {
+            assertEq(
+                uint8(l1T1ERC7683.orderStatus(helpers[i].orderId)),
+                uint8(IT1ERC7683.Status.SETTLED),
+                "Order should be settled"
+            );
         }
     }
 
@@ -112,7 +124,7 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
 
         bytes32[] memory treeLeaves = new bytes32[](intentCount);
         MerkleLeafHelper[] memory helpers = new MerkleLeafHelper[](intentCount);
-        for (uint i = 0; i < intentCount; i++) {
+        for (uint256 i = 0; i < intentCount; i++) {
             helpers[i] = _openFillOrder_and_generateMerkleLeaf(kakaroto, i % 2 == 0 ? vegeta : karpincho, amounts[i]);
             treeLeaves[i] = helpers[i].treeLeaf;
             summedAmounts += amounts[i];
@@ -122,7 +134,7 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
         originReader.commitProofOfReadRoot(batchIndex, root);
 
         bytes[] memory encodedProofs = new bytes[](intentCount);
-        for (uint i = 0; i < intentCount; i++) {
+        for (uint256 i = 0; i < intentCount; i++) {
             bytes memory flattenedProof = _flattenProof(tree.getProof(treeLeaves, i));
             encodedProofs[i] = abi.encode(batchIndex, helpers[i].requestId, i, helpers[i].result, flattenedProof);
         }
@@ -133,7 +145,10 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
         uint256 gasBefore = gasleft();
         l1T1ERC7683.handleBatchOfReadResultsWithProofs(encodedProofs);
         uint256 gasAfter = gasleft();
-        console.log("Gas used for handling proofs in test_ERC7683BatchSolverRepayment_twoSolvers_withBatching:", gasBefore - gasAfter);
+        console.log(
+            "Gas used for handling proofs in test_ERC7683BatchSolverRepayment_twoSolvers_withBatching:",
+            gasBefore - gasAfter
+        );
 
         uint256 balanceVegetaAfter = inputToken.balanceOf(address(vegeta));
         uint256 balanceKarpinchoAfter = inputToken.balanceOf(address(karpincho));
@@ -144,8 +159,12 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
             "vegeta/karpincho balance increased by batch inputs amount"
         );
 
-        for (uint i = 0; i < intentCount; i++) {
-            assertEq(uint8(l1T1ERC7683.orderStatus(helpers[i].orderId)), uint8(IT1ERC7683.Status.SETTLED), "Order should be settled");
+        for (uint256 i = 0; i < intentCount; i++) {
+            assertEq(
+                uint8(l1T1ERC7683.orderStatus(helpers[i].orderId)),
+                uint8(IT1ERC7683.Status.SETTLED),
+                "Order should be settled"
+            );
         }
     }
 
@@ -153,17 +172,16 @@ contract SolverRepaymentBatchingTest is T1XChainReaderTest {
         address opener,
         address solver,
         uint256 amount
-    ) internal returns (MerkleLeafHelper memory merkleLeafHelper) {
+    )
+        internal
+        returns (MerkleLeafHelper memory merkleLeafHelper)
+    {
         (, bytes32 orderId, bytes32 requestId) = _openAndFillOrder(opener, solver, amount);
         bytes memory result = abi.encode(l2T1ERC7683.getFilledOrderStatus(orderId));
         bytes32 xChainReadResultHash = keccak256(result);
         bytes32 treeLeaf = keccak256(abi.encodePacked(xChainReadResultHash, requestId));
-        merkleLeafHelper = MerkleLeafHelper({
-            orderId: orderId,
-            treeLeaf: treeLeaf,
-            result: result,
-            requestId: requestId
-        });
+        merkleLeafHelper =
+            MerkleLeafHelper({ orderId: orderId, treeLeaf: treeLeaf, result: result, requestId: requestId });
     }
 
     function _flattenProof(bytes32[] memory proof) internal pure returns (bytes memory proofBytes) {
